practice converting enum variants using as

+++ compound_types/enum/enum.rs
@@ -14,16 +14,16 @@ enum Number1 {
 
 // C-like enum
 enum Number2 {
-  Zero = 0.0,
-  One = 1.0,
-  Two = 2.0,
+  Zero = 0,
+  One = 1,
+  Two = 2,
 }
 
 
 fn main() {
   // An enum variant can be converted to a integer by `as`
-  assert_eq!(Number::One, Number1::One);
-  assert_eq!(Number1::One, Number2::One);
+  assert_eq!(Number::One as u8, Number1::One as u8);
+  assert_eq!(Number1::One as u8, Number2::One as u8);
 
   println!("Success!");
 }
